third_party/WebKit/LayoutTests/fast/dom/setter-type-enforcement-expected.txt

Tests type enforcement on DOM setters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body = nodelist; threw exception HierarchyRequestError: Failed to set the 'body' property on 'Document': The node provided is invalid..
PASS table.tHead = nodelist; threw exception NotFoundError: Failed to set the 'tHead' property on 'HTMLTableElement': The new child element is null..
PASS successfullyParsed is true

TEST COMPLETE
